chore: rename foo lib * chore: update lock file
