+++ examples/mfsu-e2e/foo/index.js
